Refaktorert spesialhåndtering av vilkårsresultat, oppdatert visning av delvilkår slik at Ja kommer alltid først
